[2] refactor: isolate logic in files and create custom SignalEvent
