+++ src/lib/observers/mutationListeners.ts
@@ -1,0 +1,10 @@
+export const onAttributeMutation = (
+  mutationsList: MutationRecord[],
+  callback: (mutation: MutationRecord) => void
+) => {
+  mutationsList.forEach((mutation) => {
+    if (mutation.type === 'attributes') {
+      callback(mutation);
+    }
+  });
+};
