+++ src/routes/+page.svelte
@@ -1,24 +1,24 @@
 <script lang="ts">
-  import { spy } from '$lib/action/spy.js';
+  import { spy } from '$lib/actions/spyAction.js';
+  import type { SignalEvent } from '$lib/events/SignalEvent.js';
 
   let isClicked = false;
   let test = 0;
-</script>
 
-<button
-  id="button"
-  data-test={test}
-  data-clicked={isClicked}
-  on:click={() => {
+  function onClick() {
     isClicked = !isClicked;
     test++;
-  }}
->
-  Click me
-</button>
+  }
+
+  function onSignal(e: SignalEvent) {
+    console.log(e.detail);
+  }
+</script>
+
+<button id="button" data-test={test} data-clicked={isClicked} on:click={onClick}>Click me</button>
 
 <div
-  on:signal={(e) => console.log(e.detail)}
+  on:signal={onSignal}
   use:spy={{
     target: 'button',
     intel: ['data-clicked', 'data-test'],
